detail riwayat tes

+++ app/Http/Controllers/Admin/RiwayatController.php
@@ -3,8 +3,12 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\DetailHasilMinatKarir;
+use App\Models\HasilKepribadian;
+use App\Models\HasilMinatKarir;
 use App\Models\TestHistory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class RiwayatController extends Controller
 {
@@ -15,7 +19,7 @@ class RiwayatController extends Controller
      */
     public function index()
     {
-        $data = TestHistory::get();
+        $data = TestHistory::with('user')->get();
         $title = 'Data Riwayat Tes';
         return view('admin.riwayat', compact('data'))->with('title', $title);
     }
@@ -72,7 +76,7 @@ class RiwayatController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        
     }
 
     /**
@@ -84,9 +88,68 @@ class RiwayatController extends Controller
     public function destroy(Request $request, $id)
     {
         if ($request->isMethod('post')) {
-            TestHistory::where(['id' => $id])->delete();
-            return redirect()->back()->with('toast_success', 'Hapus Data Berhasil');
+
+            if ($request->type == 'Minat Karir') {
+                $hasilMinat = HasilMinatKarir::where('test_history_id', $id)->first();
+                DetailHasilMinatKarir::where('hasil_minat_karir_id', $hasilMinat->id)->delete();
+                HasilMinatKarir::where('test_history_id', $id)->delete();
+                TestHistory::where(['id' => $id])->delete();
+            } 
+            if ($request->type == 'Gaya Kepribadian') {
+                HasilKepribadian::where('test_history_id', $id)->delete();
+                TestHistory::where(['id' => $id])->delete();
+            }
+
+            return redirect('/admin/riwayat')->with('toast_success', 'Hapus Data Berhasil');
         }
-        return redirect()->back()->with('toast_success', 'Hapus Data Berhasil');
+        return redirect('/admin/riwayat')->with('toast_success', 'Hapus Data Berhasil');
+    }
+
+    public function detailMinatKarir($id) 
+    {
+        $result = HasilMinatKarir::where('test_history_id', $id)->with(['user', 'test', 'detail'])->first(); // ambil data hasil test
+        $minatkarir = DetailHasilMinatKarir::where('hasil_minat_karir_id', $result['id'])->orderBy('point', 'desc')->get(); // ambil detail hasil
+        $data_diagram = [
+            intval($minatkarir[0]['point'] * 10),
+            intval($minatkarir[1]['point'] * 10),
+            intval($minatkarir[2]['point'] * 10),
+            intval($minatkarir[3]['point'] * 10),
+            intval($minatkarir[4]['point'] * 10),
+            intval($minatkarir[5]['point'] * 10),
+        ];
+
+        $colors = [
+            'Realistic' => '#54599e',
+            'Investigative' => '#845EC2',
+            'Artistic' => '#D65DB1',
+            'Social' => '#FF6F91',
+            'Enterprise' => '#FF9671',
+            'Conventional' => '#FFC75F',
+        ];
+
+        $startedAt = \Carbon\Carbon::parse($result->test->started_at);
+        $finishedAt = \Carbon\Carbon::parse($result->test->finished_at);
+
+        $durasisec = $startedAt->diffInSeconds($finishedAt); // ambil data durasi test dalam detik
+        $durasimin = $startedAt->diffInMinutes($finishedAt); // ambil data durasi test dalam menit
+        $result['durasi_test'] = $durasimin . ' menit ' . $durasisec % 60 . ' detik'; // data durasi
+
+        // return response()->json(['test_data' => $result, 'hasil' => $minatkarir, 'data_diagram'=>$data_diagram, 'title' => 'Hasil Tes Minat Karir | CDC Polije']);
+        return view('admin.detail-minatkarir', ['test_data' => $result, 'hasil' => $minatkarir, 'data_diagram'=>$data_diagram, 'colors' => $colors, 'title' => 'Riwayat Tes']);
+    }
+
+    public function detailKepribadian($id) 
+    {
+        $result = HasilKepribadian::where('test_history_id', $id)->with(['kepribadian','user', 'test'])->first(); // ambil data hasil test
+
+        $startedAt = \Carbon\Carbon::parse($result->test->started_at);
+        $finishedAt = \Carbon\Carbon::parse($result->test->finished_at);
+
+        $durasisec = $startedAt->diffInSeconds($finishedAt); // ambil data durasi test dalam detik
+        $durasimin = $startedAt->diffInMinutes($finishedAt); // ambil data durasi test dalam menit
+        $result['durasi_test'] = $durasimin . ' menit ' . $durasisec % 60 . ' detik'; // data durasi
+
+        return view('admin.detail-kepribadian', ['hasil' => $result, 'title' => 'Riwayat Tes', 'durasi' => $durasisec]);
+        // return response()->json( ['hasil' => $result, 'title' => 'Hasil Tes Gaya Kepribadian | CDC Polije', ]);
     }
 }
